LayoutTests/inspector/page/setEmulatedMedia-expected.txt

Tests for the Page.setEmulatedMedia command.


== Running test suite: Page.setEmulatedMedia
-- Running test case: Page.setEmulatedMedia.initial
PASS: Page should not match print media.
PASS: Body color should be blue.

-- Running test case: Page.setEmulatedMedia.print
PASS: MediaQueryList listesner fired.
PASS: Page should now match print media.
PASS: Body color should be green.

-- Running test case: Page.setEmulatedMedia.reset
PASS: MediaQueryList listesner fired.
PASS: Page should now not match print media.
PASS: Page should now match print media.
